refactor(etl): separate portable pipeline ops from Prefect orchestration
The portable core functions (compute_run_id, _compute_stats, _apply_filters,
_assign_splits, _df_to_records, _build_shards, _write_jsonl) lived inside
pipeline.py alongside Prefect task/flow decorators, making it impossible to
import or test them without a Prefect installation.

Move the portable cores to ops.py and rename pipeline.py to prefect.py so
the module boundary reflects its single responsibility. Relocate
lung_out_of_frame to processors.py where it is consumed. Update __init__.py
to re-export the full public API from the new layout.
