+++ radiologist-etl/src/radiologist/etl/__init__.py
@@ -22,4 +22,40 @@
 
 from __future__ import annotations
 
-__all__: list[str] = []
+from radiologist.etl.filters import filter_iqr, filter_lung_out_of_frame
+from radiologist.etl.manifest import JsonlWriter, ManifestRecord, ParquetWriter
+from radiologist.etl.ops import compute_run_id
+from radiologist.etl.prefect import (
+    apply_filters_task,
+    assign_splits_task,
+    build_shards_task,
+    compute_stats_task,
+    etl_flow,
+    write_jsonl_task,
+)
+from radiologist.etl.processors import StatsProcessor, lung_out_of_frame
+from radiologist.etl.shards import build_shards
+from radiologist.etl.split import assign_split
+from radiologist.etl.stats import StatExtractor, lung_asymmetry, make_haralick
+
+__all__: list[str] = [
+    "apply_filters_task",
+    "assign_splits_task",
+    "assign_split",
+    "build_shards",
+    "build_shards_task",
+    "compute_run_id",
+    "compute_stats_task",
+    "etl_flow",
+    "filter_iqr",
+    "filter_lung_out_of_frame",
+    "JsonlWriter",
+    "lung_asymmetry",
+    "lung_out_of_frame",
+    "make_haralick",
+    "ManifestRecord",
+    "ParquetWriter",
+    "StatExtractor",
+    "StatsProcessor",
+    "write_jsonl_task",
+]
